+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -59,6 +59,8 @@ class ClassConfig {
   final bool includeIfNull;
   final Map<String, String> ctorParamDefaults;
   final List<DartObject> converters;
+  final String unionDiscriminator;
+  final UnionRename unionRename;
 
   const ClassConfig({
     required this.anyMap,
@@ -75,6 +77,8 @@ class ClassConfig {
     required this.genericArgumentFactories,
     required this.ignoreUnannotated,
     required this.includeIfNull,
+    required this.unionDiscriminator,
+    required this.unionRename,
     this.converters = const [],
     this.ctorParamDefaults = const {},
   });
@@ -105,6 +109,9 @@ class ClassConfig {
         fieldRename: config.fieldRename ?? ClassConfig.defaults.fieldRename,
         disallowUnrecognizedKeys: config.disallowUnrecognizedKeys ??
             ClassConfig.defaults.disallowUnrecognizedKeys,
+        unionDiscriminator: config.unionDiscriminator ??
+            ClassConfig.defaults.unionDiscriminator,
+        unionRename: config.unionRename ?? ClassConfig.defaults.unionRename,
         // TODO typeConverters = []
       );
 
@@ -125,6 +132,8 @@ class ClassConfig {
     genericArgumentFactories: false,
     ignoreUnannotated: false,
     includeIfNull: true,
+    unionDiscriminator: 'type',
+    unionRename: UnionRename.none,
   );
 
   JsonSerializable toJsonSerializable() => JsonSerializable(
@@ -142,6 +151,8 @@ class ClassConfig {
         genericArgumentFactories: genericArgumentFactories,
         fieldRename: fieldRename,
         disallowUnrecognizedKeys: disallowUnrecognizedKeys,
+        unionDiscriminator: unionDiscriminator,
+        unionRename: unionRename,
         // TODO typeConverters = []
       );
 }